+++ src/main/java/hiyen/Main.java
@@ -1,8 +1,12 @@
 package hiyen;
 
+import hiyen.product.Product;
+import hiyen.ui.Console;
+import hiyen.ui.PrintConsole;
 import java.util.Scanner;
 
 public class Main {
+
     public static void main(String[] args) {
         VendingMachine vendingMachine = new VendingMachine();
         Scanner scanner = new Scanner(System.in);
