+++ src/main/java/hiyen/product/Cider.java
@@ -1,6 +1,7 @@
-package hiyen;
+package hiyen.product;
 
 public class Cider extends Drink {
+
     public Cider(final String name, final int price) {
         super(name, price);
     }
