+++ src/main/java/hiyen/product/Coke.java
@@ -1,6 +1,7 @@
-package hiyen;
+package hiyen.product;
 
 public class Coke extends Drink {
+
     public Coke(final String name, final int price) {
         super(name, price);
     }
