+++ src/main/java/hiyen/product/DoctorPepper.java
@@ -1,6 +1,7 @@
-package hiyen;
+package hiyen.product;
 
 public class DoctorPepper extends Drink {
+
     public DoctorPepper(final String name, final int price) {
         super(name, price);
     }
