+++ src/main/java/hiyen/product/Product.java
@@ -1,7 +1,10 @@
-package hiyen;
+package hiyen.product;
 
 public interface Product {
+
     String getName();
+
     int getPrice();
+
     String consume();
 }
