+++ src/main/java/hiyen/product/Twix.java
@@ -1,6 +1,7 @@
-package hiyen;
+package hiyen.product;
 
 public class Twix extends Food {
+
     public Twix(final String name, final int price) {
         super(name, price);
     }
