+++ src/main/java/hiyen/ui/Console.java
@@ -1,6 +1,8 @@
-package hiyen;
+package hiyen.ui;
 
 public interface Console {
+
     String read();
+
     void print(String message);
 }
