+++ src/main/java/hiyen/ui/PrintConsole.java
@@ -1,8 +1,9 @@
-package hiyen;
+package hiyen.ui;
 
 import java.util.Scanner;
 
 public class PrintConsole implements Console {
+
     private final Scanner scanner;
 
     public PrintConsole(final Scanner scanner) {
